Modify error handling in join_role_stats and move it to the Information Cog
