+++ bot/utils/helpers.py
@@ -1,8 +1,7 @@
 from abc import ABCMeta
-from typing import Dict, List, Optional
+from typing import Optional
 
-from discord import Guild
-from discord.ext.commands import BadArgument, CogMeta
+from discord.ext.commands import CogMeta
 
 
 class CogABCMeta(CogMeta, ABCMeta):
@@ -31,14 +30,3 @@ def has_lines(string: str, count: int) -> bool:
 def pad_base64(data: str) -> str:
     """Return base64 `data` with padding characters to ensure its length is a multiple of 4."""
     return data + "=" * (-len(data) % 4)
-
-
-def join_role_stats(role_ids: List[int], name: str, guild: Guild) -> Dict[str, int]:
-    """Return a dict object with the number of `members` of each role given, and the `name` for this joined group."""
-    members = []
-    for role_id in role_ids:
-        if (role := guild.get_role(role_id)) is None:
-            raise BadArgument("Unable to fetch role data, the specified role does not exist.")
-        else:
-            members += role.members
-    return {name: len(set(members))}
